requirements.txt
# Repository : bigip-icontrol-rce-research
# Path       : requirements.txt
# Purpose    : Pins runtime Python dependencies for all services.
# Layer      : config
# SDLC Phase : implementation
# ASVS Ref   : V14.2.1
# OWASP Ref  : A06
# Modified   : 2026-04-11
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
fastapi==0.110.1
uvicorn==0.29.0
sqlalchemy==2.0.29
pydantic==2.6.4
httpx==0.27.0
